+++ src/components/ui/detectionItem/detectionItem.js
@@ -65,7 +65,7 @@ export function detectionItem(detectionItems) {
           </div>
         </div>
       <div class="configItem-action-container">
-        <button class="button-primary button-reset button_deleteDetection">
+        <button class="button-primary button-reset button_deleteDetection" title="Löschen">
           <?xml version="1.0" encoding="utf-8"?>
           <!-- Designed by Empty Soul  -->
           <svg version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px"